0.5.0 - show list's element definition always (if not empty)
