+++ example/issue15a.conf
@@ -1,0 +1,4 @@
+dd: "double"
+foo: {
+  bar: string
+}
