+++ example/issue15b.conf
@@ -1,0 +1,2 @@
+names: [ string ]
+ages: [ integer ]
