+++ example/issue15c.conf
@@ -1,0 +1,6 @@
+positions: [
+  {
+    lon: "double"
+    lat: "double"
+  }
+]
