+++ example/issue15d.conf
@@ -1,0 +1,4 @@
+baz: [ [ {
+  dd: double
+  aa: "boolean?"
+} ] ]
